+++ kz-pompei-hotconfig-etcd/src/kz/pompei/hotconfig/etcd/ConfigTunnelEtcdBuilder.java
@@ -8,12 +8,23 @@ public class ConfigTunnelEtcdBuilder {
   private          Client client;
   private @NonNull String keyPrefix       = "/kz-pompei-conf-etcd/";
   private @NonNull String noticeExtension = ".notice";
+  private @NonNull String errorPrefix     = "ERROR ";
 
   /**
    * Extension appended to the configuration key when storing notice lines.
    */
   public @NonNull ConfigTunnelEtcdBuilder noticeExtension(@NonNull String noticeExtension) {
     this.noticeExtension = noticeExtension;
+    return this;
+  }
+
+  /**
+   * Prefix used after the leading {@code #} marker for parameter error lines.
+   * <p>
+   * For example, the default value {@code ERROR } is written as {@code #ERROR } in etcd payloads.
+   */
+  public @NonNull ConfigTunnelEtcdBuilder errorPrefix(@NonNull String errorPrefix) {
+    this.errorPrefix = errorPrefix;
     return this;
   }
 
@@ -45,10 +56,14 @@ public class ConfigTunnelEtcdBuilder {
     return noticeExtension;
   }
 
+  public @NonNull String errorPrefix() {
+    return errorPrefix;
+  }
+
   public ConfigTunnelEtcd build() {
     if (client == null) {
       throw new IllegalArgumentException("Hf3Lp9Qw2E :: `client` must be specified");
     }
-    return new ConfigTunnelEtcd(new ConfigTunnelEtcd.Def(client, keyPrefix, noticeExtension));
+    return new ConfigTunnelEtcd(new ConfigTunnelEtcd.Def(client, keyPrefix, noticeExtension, errorPrefix));
   }
 }